openstack-ardana-gating: full tempest and entry-scale-kvm
This change replace the `std-min` and `std-3cp` input models
currently employed by the Ardana gating jobs with an
entry-scale-kvm input model and replaces the 'ci' tempest
filter with a full set of tempest filters.

With the recent cleanup of Ardana tempest filters and the
efforts to stabilize the CI, we can finally have a full tempest
coverage in the gate. The `entry-scale-kvm` input model is
also better suited for a gating job, because it is closer to
what customers use.

This change also updates all other jobs to reflect the new
tempest filters and run only those that are applicable [1].

[1] https://confluence.suse.de/display/qualityperformance/Tempest+Test+Coverage+on+Ardana