openstack-ardana: override OpenStack service worker count (SCRD-7496)
The number of API and RPC workers configured for various OpenStack
services in Ardana is dynamically calculated based on the number
of CPUs. However, the formula is currently incorrect and leads
to a huge number of processes being created on the virtual Ardana
cloud installations - there are 4 times as many workers as there
should be - which has negative side effects on the overall
performance.

To make a simple comparison, the general formula used by Ardana to
calculate the number of workers is:

  min(ansible_processor_count * ansible_processor_cores * 2, max_workers)

where `max_workers` has a value of `10` for most services, but sometimes
can be `20` and sometimes `8`, whereas the OpenStack Ansible formula is:

  min(ansible_processor_vcpus / 2, 16)

and the general rule used for Crowbar is:

  min(max(node["cpu"]["total"], 2), 4)

This change overrides the worker count by passing a number of extra
ansible variables when calling Ardana playbooks such as `site.yml`
and `ardana-update.yml`. The values correspond to the current flavor
used for controller nodes, which has 4 vCPUs.

NOTE: this measure is hopefully temporary. It can be removed if
and when the formula used to calculate the number of workers is
corrected in Ardana. However, the number of changes required for that
is extensive (one for each service), so this approach will be used
instead for the time being.