Blacklist known to fail neutron_tempest_plugin tests (SCRD-8681)
As part of enabling the neutron_vpnaas tempest plugin for OpenStack
Rocky we need to ensure that the python-neutron-tempest-plugin package
is installed, which leads to tempest failures during a cleanvm run
for the tests patching the pattern:

    neutron_tempest_plugin\.api\.admin\.test_tag\
      \.Tag(Filter|)(QosPolicy|Trunk)TestJSON

This patch adds a general purpose blacklisting mechanism for known to
fail tests to qa_openstack.sh, stored in qa_openstack_blacklist.txt
under /var/lib/openstack-tempest-test, which is passed to the relevant
commands when they are invoked.

The known to fail tests identified above are appended to the blacklist
file when the cloudsource is 'openstackrocky' when the Neutron tempest
plugin package is detected as being installed, ensuring that they will
not cause tempest run failures.