openstack-ardana-gerrit: versioned configuration
The configuration used by Ardana Gerrit and trackupstream Jenkins
jobs and scripts is different from one cloud version to the next,
so it needs to be versioned.

This commit adds a new version dimension to the central configuration
file that controls all of the above, and also changes all other scripts
and Jenkins job definitions to use this new dimension.