cloud-deployment: List+Diff installed packages
Simple functionality(based on Victor's afforts) which can be called when
needed to collect(list) installed packages on the cloud nodes and|or to diff
changes(installed packages) before and after a update for example. Works on Ardana
and Crowbar.

When called with "list" it collects installed packages, when with "diff", then
it run list with "after" state and do the diff with previous state. States can
be defined as shown below or default values will be used.

Includes:

* role
* playbook
* pipelines changes in templates for ardana & crowbar
* enhancement of bootstrap-vcloud-nodes.yml to get ssh_pub_keys to ardana cloud
nodes for root and ardana uses