openstack-ardana: mariadb performance tunings (SCRD-7496)
When mariadb is running on top of shared (ceph) storage with the
default configuration values, there is a huge impact on its
performance, because database commit operations must wait until
disk flush operations are complete (and with shared storage,
that takes longer).

There are two solutions to this:

1. move the database to the root filesystem, which is not
currently implemented as a ceph volume (the ephemeral volume
in the ECP is still set up as a local volume on the compute node)

2. decouple the flush and commit operations, by tweaking the
innodb configuration (i.e. set the `innodb_flush_log_at_trx_commit`
configuration option [1] to 2). This option requires matching changes
in the `db-ansible` repository [2].

This commit implements both these options and adds a top-level
`ardana_dbmq_use_root_volume` ansible variable which can be used to
switch between them. The default choice is to use local storage,
because this doesn't depend on Ardana ansible package changes and
can be enabled immediately.

[1] https://dev.mysql.com/doc/refman/8.0/en/innodb-parameters.html#sysvar_innodb_flush_log_at_trx_commit
[2] https://gerrit.suse.provo.cloud//#/c/5957