openstack-ardana: use bigger SES volume for nova
When the cinder LVM back-end is not enabled, the compute node
doesn't need the additional volume mounted to /var/lib/nova.
Instead, this additional storage can be added to the SES cluster.

This helps with running tempest, because it allows more workers
to run in parallel and create VMs without running into storage
limit.