Fix magnum tempest testsetup
magnum jobs on Cloud9+ fail with:

The specified regex doesn't match with anything

which is because the tests have been moved out of tree into a separate
magnum_tempest tests package. adjust regexp for rocky only, which
means we need to make it configurable and use the old values. Also
drop the aarch64 part because PM asked us to deinvest in aarch64 for
cloud8.