try neutron proposal for controller, if nova does not exist
currently without nova, neutron and designate cannot be deployed, this is
a pretty hard requirement which does not really exist in the product.

One can have keystone and neutron working together to test the
integration, but having nova (glance, cinder) simply increases
deployment time as well as chef-client run time which makes
testing quick changes to the recipes painfully difficult

This patch just make sure novacontroller variable is pointed to the
correct host by looking at the neutron-server attribute of the neutron
proposal. This could be extended to other barclamps, but that is not the
aim of the patch.

The variable is not really a "nova controller" but just a simple
controller and it begs being renamed everywhere, but since the
variable has always been called novacontroller, i dare not change that.