Update /etc/hosts on controller nodes
In a deployment where we specify an external
name for the public endpoints, the /etc/hosts
file need to be updated with this entry. Earlier,
we did this on only deployer node.
However, services like heat communicate to
keystone via the public endpoint. This requires
that the /etc/hosts file on the controller nodes
need to be updated as well.