input-model-generator: move default route to CONF network
Move the default route to the `CONF` network in all the scenarios
that use the `standard` network template (`standard`, `std-lmm`,
`std-split`). This has the effect that the default route is set
up on the same interface used by the lifecycle manager, which is
the same interface that is associated with the floating IP on
the deployer node. When IPv6 is enabled, the default route will
be IPv4, which allows the nodes to communicate with the outside
world.