mkcloud: Allow supplying a network.json patch
qa_crowbarsetup already has support for patching
/etc/crowbar/network.json before deploying the admin node. However
currently that patch needed to be manually uploaded to the node. This
patch copies any "network.json.patch" that exists in the cwd into the
admin node (similar to what is already done for "crowbar.patch" and
applies it to /etc/crowbar/network.json.