ironic: Improve virtualbmc startup
Sometimes `vbmc start` returns "BMC instance ... already running"
on first attempt and/or if run to quickly after `vbmc add`.
In such cases, the BMC needs to be re-created before it will start
correctly.

Wrapping the (re)start logic in a function makes it easier to use
it in `wait_for`.