ipv6: Configure radvd if required during prepare
On hosts that are not mkch* hosts, we may need to have radvd running
so the admin node can use SLAAC to get an address. If the mkcloud driver
is libvirt and if radvd isn't running, start it, first adding some
configuration if required.

To help with keeping the configuration DRY, refactor the existing
generate-radvd-conf to use the same template.