socok8s: Only test PRs from admins or users with write permissions
From the docs:

From users with Admin or Write permission
Pull requests forks will be treated as trusted if and only if the fork
owner has either Admin or Write permissions on the origin repository.
This is the recommended policy. Note that this strategy requires the
Review a user's permission level API, as a result on GitHub Enterprise
Server versions before 2.12 this is the same as trusting Nobody.