mkcloud: improve image cache usage
In order to prevent the image cache directory to fill up
the disk the cache should share its images with all jobs.
A 'per' job cache does not make sense for an image thats
just used once.

So the value of $cache_dir_default is now a shared directory
and no longer contains the cloud name.
To not break running deployments rsync needs to drop the
'inplace' parameter, so that it first syncs the new image
and then does a rename at the end.
This guarantees that
1) there is always one consistent image to deploy from,
2) no running jobs read from currupt files,
3) the cache really caches for all jobs on the host.

The compress parameter can also be dropped because the image
files are already compressed. So we can save CPU cycles.