openstack-cloud-heat: separate jobs for engcloud and susecloud
Create one dedicated `openstack-cloud-heat` Jenkins job for each
of the available OpenStack cloud platforms that can be used in the
CI to deploy heat stacks.

Having one job running exclusively for a target OpenStack platform
means it can be used to monitor the platforms individually for history
of failures and overall reliability.