designate: create a test port and nslookup it
Added a basic test for designate recordset being created when a
floating ip is assigned to a port on a fixed network.

ping will not work because the port is still down as it is not assigned
to a vm. This helps test designate without
nova/glance/cinder/milkyway/...  and other astronomical services.

Also fixed a type in pools file name.