Add upgradecloud9 cloudsource (SOC-9340)
This adds the temporary Devel:Cloud:9:Upgrade project as a
cloudsource. This project is meant for quick turnaround and
parallelization on fixing the cloud-mkcloud9-job-upgrade-disruptive
CI job: one person can add quick and dirty patches to this
repository until the job is green while others can polish
the existing patches and get them upstream in parallel.