Add periodic socok8s airship from rpm job
This tests the socok8s project from RPM rather than from source. As we
are testing an RPM, rather than installing it onto the host worker,
install into and operate it from a podman container.

This job requires the jenkins worker to be set up with podman. This has
already been done manually for the cloud-ccp-ci node.