Use git ref location that prevents overlap
Someone created a branch named pr/282 which with the git fetch config
creates a path for the local remotes ref where the PR of that
number and the branch of that name overlap. This prevents fetching all
updates in the git repo.

cloud-specs$ git fetch origin
fatal: Cannot fetch both refs/heads/pr/282 and refs/pull/282/head to
 refs/remotes/origin/pr/282

Afterwards the cloud-specs CI job failed to checkout the PR to test as
it was not locally availabe and thus tested master instead.

Fix this by chosing a config where those don't overlap.

Also set error on exit to fail CI instead of silently passing.