openstack-cloud-heat: parameterized timeout
The timeout used by heat when creating/deleting a heat stack is
configurable. The OpenStack heat API call (and consequently
the wrapping ansible playbook) will fail and exit if the heat
stack create or delete operation exceeds this timeout value,
and will not print any output during this time.

The Jenkins pipeline also uses a second, inactivity timeout
value (based on output lines being added to the Jenkins log),
in case something else goes wrong. The Jenkins job inactivity timeout
needs to account for the heat stack timeout (i.e. the value must
a little higher).

All this logic is in reality based on a single parameter - the heat
stack timeout - but the implementation so far used two independent
values. To add to the complexity, the supported ECP and SUSE cloud
platforms differ in performance (the SUSE cloud takes roughly twice
as much time to create a heat stack).

This commit makes things easier to maintain, by adding a hidden
parameter to the openstack-cloud-heat jobs, which controls both
timeout values.