Refactoring: unified and improved heat stack naming scheme (SOC-8475)
Change the naming scheme used for the generated heat stacks and
their resources to reflect the fact that they are used in common
for both Crowbar and Ardana and to be better aligned with the
socok8s heat stack naming scheme.

A virtual cloud deployment usually consists of two heat stacks,
one for the actual cloud setup and the other one for the SES
cluster used for storage. This is the new naming scheme, based
on the `cloud_env` parameter value supplied by the user or the
periodic Jenkins jobs, which uniquely identifies a cloud deployment:

  - the cloud heat stack name will be `<cloud_env>-cloud` and the
    names of its resources will also be prefixed by the same identifier

  - the SES heat stack name will be `<cloud_env>-ses` and the
    names of its resources will also be prefixed by the same identifier

The heat_stack ansible role will continue to support the old
naming scheme, to ensure that the existing heat stacks that were created
using the old naming scheme are properly cleaned up.