openstack-ardana: add option to update services serially
Although the documentation recommends running service updates serially,
for the CI we have been running service updates in parallel due to the
long time it takes to complete when running in serial (5h vs 9h).

However, during the testing of MUs for cloud 8, the update failed when
running in parallel (SOC-9524) and an option to run services update in
serial was needed.