openstack-ardana: make packages from extra-repos available to all nodes
Instead of adding the extra repositories on the deployer node as a
zypper repository, create a new repository on the deployer and make
all RPMs from the extra repositories available to all cloud nodes by
setting up this repository on all nodes.

This new repository will be configure with higher priority, meaning that
all packages from the `extra_repos` will be installed even if there are
newer packages available on other repositories.

https://jira.suse.com/browse/SOC-7800