heat-generator: fix heat stack output
The stack output is including the deployer/admin node on the list of
controller-conf-ips. This change is fixing that by updating the list of
CLM service components (octavia/manila-client was massing) and fixing a
conditional check on the generate_heal_model module.

This change also removed the `standard.yml` file as it is not being
used.