gitlint: Allow trivial changes to be without reference (SOC-9559)
The author of a commit should make a conscious decision to either add a
reference or not. Adding trivial instead of bsc or SOC enables this.
This change prevents that adding a refernce is forgotten as the
syntax requires to think about it.

trivial -> simple unplanned changes
noref -> more complex unplanned changes
typo -> typo fixes only

See
https://github.com/SUSE-Cloud/automation/pull/3509#discussion_r298963183
for more details.