openstack-maintenance-gating: add crowbar deploy jobs (SOC-9599,SOC-9669)
Add Crowbar jobs to the set of jobs executed for a maintenance
update.

This commit switches the MU gating job implementation to use
a configuration yaml file listing the jobs and the parameters
that need to be launched for each cloud version rather than
hard-coding them in the groovy library.

The commit also enables serializing the service update process
for cloud nodes in Ardana deploy-and-update MU jobs.