crowbar: run tempest using filters
This change adds a playbook/role for running tempest on crowbar
deployments using filters (which are based on services) in a similar way
to how it is done in Ardana.

This change also adds support for re-running failed tests serially by
setting `tempest_retry_failed` to true (defaults to false).