Add prepareupgraderepos step
This is just a small feature for simplifying manual testing of upgrade.
In normal upgrade scenarios, repos need to be prepared and updated by the
user. Exposing the "prepare" functions already present in mkcloud enables
developers to use these instead of various hacks (e.g. local modifications
to mkcloud) that were done in the past.