iscsictl.py: fix SBD persistence issues (SOC-9821)
Rebooting the admin node in an SBD enabled HA Crowbar setup had
issues with the shared iSCSI volume, because the iSCSI configuration
wasn't persisted on the admin node.

The `tcm_dump --b` parameter expects an value of 1 if the default
configuration is to be overwritten [1]. Our `iscsictl.py` script
was using an incorrect value of `OVERWRITE`, which had no effect.

However, supplying `tcm_dump --b=1` overwrites the `/etc/target/tcm_start.sh`
and `/etc/target/lio_start.sh` files [2], which the `ilo-utils` package
pre-installs with some useful commands such as loading required modules,
so we can't use that either.

As it turns out, the `target.service` unit already has support for
persisting the configuration [3]. All we have to do is restart the
service after completing the configuration.

[1] https://github.com/Datera/lio-utils/blob/d488da31b062bb374afdc8ae9afa2de7f044a3b6/tcm-py/tcm_dump.py#L320
[2] https://github.com/Datera/lio-utils/blob/d488da31b062bb374afdc8ae9afa2de7f044a3b6/tcm-py/tcm_dump.py#L298-L299
[3] https://build.suse.de/package/view_file/SUSE:SLE-12-SP3:Update/lio-utils/0030-target.service-Fixup-service-file.patch?expand=1