Introduce new `bootstrap_network` flag
For IPv6 deployments we need a separate network interface on IPv4 that
can bind to the floating IP address from Neutron and serves as the
entry and exit point to the deployment (since the ECP can not route
IPv6 to and from external addresses). This change introduces a new
setting to enable the creation of a separate network interface.

Signed-off-by: Nicolas Bock <nicolas.bock@suse.com>