iscsictl.py: setup loop device before target service start (SOC-9821)
There is no clear execution order dependency between
`/etc/rc.d/boot.local` and the startup of the target service,
which leads to situations in which the target service starts
before the loop device has had a chance to be properly set up.

Moving the setup of the loop device to the target service startup
script ensures that the commands setting up the ISCSI device used
for SBD are executed in the correct order.