qa_crowbarsetup: Use separate glance scrubber log (SOC-9929)
When looking at the scrubber.log we also search for errors unrelated to
this glance-scrubber run. As the glance-scrubber service runs regularly
it can happen that e.g. keystone is not available at a previous run. In
the gate glance-scrubber regularly runs when apache is restarted so a
ERROR is reported.