Report known jenkins job failures to Jira
Keeping track of the known issues related to failures on the cloud
jenkins jobs is kinda hard (nowadays there are many jenkins jobs
crowbar/ardana/7/8/9, sometimes it is not easy to reproduce the
issue and we think the issue was fixed, there is already a ticket opened
for the issue and another one is created...).

This change makes keeping track of those failures easier by
automatically commenting on the existing jira card whenever a know issue
happens. It also edit the job description with the link of the known
jira cards related to job failure (if there is any).