Add socok8s-trackupstream job
Similar to openstack-trackupstream and cloud-trackupstream, a new job
called soco-trackupstream is added.
This job will periodically update the Cloud:socok8s:master/socok8s RPM
package from the git source defined in the _service file.
This is needed to get an updated ISO with the latest socok8s code on a
daily base.