openstack-ardana: Set monasca tuning config to small (SOC-10145)
The monasca role chooses the tuning setting based on how much memory the
system has, meaning that for virtual deployments it is using the `demo`
setting [1]. However `monasca-thresh` is occasionally hanging during
start/stop and there is a suspicion that his happens due to the lack of
memory allocated to it.

This change override the monasca tuning parameter to use the `small`
setting. This will enable storm to use more memory (from 64 to 128 MB)
and probably fixes SOC-10145. Previous tests were made and SOC-10145 was
not reproducible when using the `small` setting.

1. http://git.suse.provo.cloud/cgit/ardana/monasca-ansible/tree/roles/monasca-variables/vars/configuration.yml#n48