mkcloud: fix timeout waiting for admin node reboot
This commit increases the netcat timeout when waiting for the
SSH daemon on the admin node to vanish. This mechanism is used
to detect whether the Crowbar admin node has already gone down
for reboot. Under adverse circumstances (slow mkcloud host,
bad timing), an update of the openssh package could trigger
the previous timeout of 1s early (way before the reboot), thus
leaving the cloud in an inconsistent state.