revert osc release process (SOC-10025)
This makes Rocky and Stein use the old process again.

OBS regularly causes conflicts that it could resolve automatically.
While each of these is a case where in the old process we would accept a
version of a package that was changed after testing started, we decided
to revert for now and revist later towards a more long term solution to
our staging needs.