qa_openstack: Do not force install quickstart and tempest-test
When calling

zypper --non-interactive -n install --force openstack-quickstart \
    openstack-tempest-test

the installation fails when both packages are in different
repositories. The failure is:

Forcing installation of 'openstack-quickstart-2019.2' from repository 'cloud'.
Package 'openstack-tempest-test' not found.

But openstack-tempest-test is provided by openstack-tempest from the
upstreamcloud repository.

The installation works fine if the "--force" option is not used.

Note: Filled also a zypper bug for it:
https://github.com/openSUSE/zypper/issues/289