Split SSH key generation and authorized_keys update
Use separate ansible actions to ensure that the key has been created
and to append the key to .ssh/authorized_keys file. Since the pub keys
are already being retrieved from the deployer node to use later when
bootstrapping the other nodes, we can use a similar lineinfile edit
action to ensure that we only add the pub key once.

This fixes an issue seen in manual mode deployments using the command
line; the issue doesn't appear to affect the Jenkins driven jobs.