Enable neutron service disablement (SOC-9780)
One of the goals of the SOC/CLM 8 to 9 upgrade tooling was to enable
testing of minimal deployments, such as a keystone+horizon only single
controller node setup.

Howver defining an extry-scale-kvm or standard scenario cloud without
neutron would fail throwing exceptions in the heat model generator
due to attempting to dereference an item with a value of None as either
a dictionary or a list; adding guard checks to skip such elements when
converting lists to dictionaries avoids the exceptions.

The reason that the exceptions were occurring was due to the fact that
if you specified compute nodes in your model, but disabled all of the
service components running on them, you ended up with compute nodes
without empty service component lists, which caused the control plane
templating to generate 'resources:' entry with no content, meaning a
value of None. This also leads to the ardana-configuration-processor
report errors for the input model. This can be avoided by only emitting
the service_group_type header line if there are actually nodes in the
associated group.

Additionally if you disable the designate, neutron, or octavia services
we still emitted references to the configuration data objects, which
leads to the configuration processor reporting warnings for unused refs
that can be avoided by making the config_data item entries conditional
on the service not being disabled when rendering the control_plane.yml.

Similarly when rendering the network_groups.yml, only generate the
designate, NEUTRON VLAN or Neutron physnet elements if the associated
service is not disabled.

Enable environment variable overrides.

Add support for using the NO_CONFIRM and NO_MITOGEN environment variables
to disable the interactive confirmation of input paramaters and whether
mitogen is configured/enabled.

The disablement of the interactive confirmation makes it easier to run
testing scenarios using the command line tooling from scripts.

Being able to disable mitogen configuration/enablement without needing
to edit the standard scripts is helpful when debugging issues that may
potentially be related to the manual tooling's use of mitogen which is
not used by the Jenkins CI tooling due to issues seen in the past.