Octavia: Add octavia next steps (SOC-10685)
The octavia barclamp installs and setups octavia, inluding creating a
octavia management network that is used for the octavia worker and
healthcheck daemons to manage amphora's.

But there are some missing steps to finalise that is left to the
operatior. This includes:

  - Having the octavia backend servers able to route to the neutron
    octavia management network.

  - Generate the SSL certs for communication between octavia workers and
    the amphora's.

The patch sets up the routing by connecting the lb_mgmt_net to the
router-floating router. Then adds a route on the octavia-backend
nodes.

The certs are generated by some new scripts in `scipts/lib/octavia`.
The first `build_certificates.sh` builds the certs, using secret of
`foobar`. Then `distribute_certificates.sh` will push them to the
octavia-backend nodes.

This is only the first part of the linked jira card. Once these next
steps are finalised, another patch or maybe this one will add octavia
as a CI/CD job.