Add Train and remove Master from trackupstream
Since all of the service packages are now maintained in rpm-packaging
and updated in Cloud:OpenStack:Upstream:Master with the
openstack-rpm-packaging-update-Master job, we don't need to run
trackupstream for Cloud:OpenStack:Master, and in fact we can't any more
because those packages don't contain _service files. However, very soon
we will need to start running trackupstream for the Train branch, so
this change just s/Master/Train:Staging/ since all the conditions that
apply to the current master branch will also apply to stable/train. The
Train part of the trackupstream matrix won't work until the
Cloud:OpenStack:Train:Staging project is created in OBS.