Setup Pool & Updates tags for Devel and Staging repos (SOC-9780)
After rsync'ing the Devel or Staging repos to the deployer, setup the
appropriate Pool & Updates tags so that any update or upgrade actions
can use them without having to disable repos checks, and eliminates
the need to remove and re-add these repo after adding the Updates tag
during the update workflow.

Ensure that the repos_to_add list is reset to an empty list at the
top of the setup_sles_cloud_repos.yml task list as otherwise, if the
task list is included multiple times in a playbook run, subsequent
inclusions will inherit the value from the previous run which may not
be desirable if used as part of setting up a new cloudsource during
upgrade.

Move some facts used exclusively in the setup_zypper_repos role to be
managed in that role's defaults/main.yml, rather than as part of the
group_vars/all hierarchy.