Enhance zypper repo management (SOC-9780)
Add support to the setup_zypper_repos role to allow it to completely
manage the lifecycle of repos, rather than just simply add them to
the system. This permits playbooks to:
  * add repos in a disabled state
  * disabled already defined repos
  * remove defined repos
  * delete the on-disk contents of downloaded repos or unmount NFS
    imported repos

Three new settings have been introduced that are used to manage the
state of zypper repos:
  * zypper_repo_enabled:
    - yes/no value indicates if repos are enabled or disabled
  * zypper_repo_state:
    - present/absent indicates if repos are added/remove
    - only valid if zypper_repo_enabled == no
  * zypper_repo_delete:
    - boolean indicating if repo on-disk contents or persistent NFS
      mount should be removed.
    - only valid if zypper_repo_state == absent

This will be useful to the Ardana/CLM upgrade process which will need
to dist-upgrade the deployer node from SP3/Cloud8 to SP4/Cloud9.

Renamed repos_to_add to repos_to_manage as a better logical fix for the
evolved usage of the fact.