Created unified gating jobs (SOC-10029, SOC-10398)
This should functially move all gating to the ECP based CI.
Changes:
1: enable Staging to Dev propigation for Cloud 8, 9 through ECP gate jobs
2: add crowbar jobs to cloud-{8,9}-gating
3: remove crowbar blacklist and ardana white list from ECP

Based code in: jenkins/ci.suse.de/openstack-submit-project.yaml

ISO modified time at the start time of the cloud-{8,9}-gating jobs that control
which packages will be promoted and since that will be the same
for crowbar and ardana, things should be as correct as they are now.
Possible more so, since the current code allows to venv rpms to trail
the RPMs in the ISOs.

Issues:

1: The ISO build times are different for Crowbar and Ardana (for the current
builds, currently a 3 hour difference in both cloud 8 and 9)

2: There is no guarantee that the Crowbar ISO will be built and will be
available in Provo before the Ardana ISO (for the current builds, in
cloud 8, the Ardana ISO was built first, whereas in cloud 9, the
Crowbar ISO was built first). This, when (and if) coupled with the above,
means there's a fair chance that the gating job will be promoting
untested Crowbar changes. This can be very bad if occurring frequently.

3: There is no guarantee that the Ardana ISO will be rebuilt every time
there's a Crowbar change (e.g. a new Ardana ISO may not be rebuilt
every time there's a crowbar-openstack package version). In other words,
the gating job will only be triggered for Crowbar changes
if they also impact Ardana. This isn't so bad, but in some cases it
will delay promoting Crowbar changes unnecessarily.