SOC-10029/SOC-10398 Created unified gating jobs
This should functially move all gating to the ECP based CI.
Changes:
1: enable Staging to Dev propigation for Cloud 8, 9 through ECP gate jobs
2: add crowbar jobs to cloud-{8,9}-gating
3: remove crowbar blacklist and ardana white list from ECP

Based code in: jenkins/ci.suse.de/openstack-submit-project.yaml

ISO modified time at the start time of the cloud-{8,9}-gating jobs that control
which packages will be promoted and since that will be the same
for crowbar and ardana, things should be as correct as they are now.
Possible more so, since the current code allows to venv rpms to trail
the RPMs in the ISOs.