Implement initial Ardana upgrade workflow (SOC-9780)
Add a new role, ardana_upgrade, loosely based on the ardana_update role
which implements the Ardana upgrade workflow.

Note that the upgrade workflow is memory intensive so we are targetting
running it against the mid-scale-kvm scenario model which spreads out
the various services across multiple nodes; the entry-scale-kvm or
standard scenario models will likely be unstable for this reason,
though the entry-scale-kvm-mml model may also work.

Additionally add 2 new top-level playbooks which are used by the Ardana
upgrade workflow:

  * ardana-disable-repos.yml
    - Leverages recent enhancements to the setup_zypper_repos role that
      allow the existing repos on a deployer node to be disabled, which
      is used by the upgrade workflow to disable the repos associated
      with the active cloudsource.

  * ardana-upgrade.yml
    - Upgrades the deployer to the active cloudsource by adding/enabling
      the cloudsource repos to the deployer
    - runs zypper dist-upgrade to upgrade the system to the (new)
      cloudsource.
    - installs/updates the Ardana cloud pattern and runs ardana-init to
      update the deployer's ardana account to the (new) cloudsource
      provided content.
    - addresses mandatory deprecations in the cloud model.
    - runs the configuration processor and regenerates the scratch
      area
    - runs the ardana-upgrade.yml playbook to upgrade the
      whole cloud to the (new) cloudsource.
    - reboots all the nodes to pick up new kernels and/or core services.

Update the openstack-ardana.Jenkinsfile to add the Upgrade stage that
calls the relevant play books with appropriate arguments. Also add the
new upgrade_cloudsource parameter definition to the pipeline template
in cloud-ardana-pipeline-template.yaml.

Ensure that cobbler-deploy.yml is called for update runs if it is a
physical or PXE boot enabled deploy.

Ensure that the bootstrap_ardana/tasks/download_cobbler_requires.yml
will download the SLE 12 SP4 ISO if we are running an upgrade scenario.

Update the flavours used for the mid-scale-kvm model to sizes that
ensure enough RAM/vCPU resources are available to reliably run the
upgrade workflow.

Enhanced the ardana_reboot role to reboot nodes in the this order:
  * deployer node
  * non-deployer controller nodes (if any)
  * compute (NOV-CMP) nodes (if any)

This ensures that we reboot compute nodes last, even if they happen
to appear in the resources earlier.

Implement a script, cloud9-input-model-upgrade, in the ardana_upgrade
role's files tree, that handles upgrading a Cloud8 input model to
Cloud9, applying any mandatory service component deprecations.

Add support to the rocketchat_notify role for the new upgrade related
highlevel playbooks, ardana-disable-repos.yml and ardana-upgrade.yml.

Added a new top-level playbook, ardana-reboot.yml, which can be used
to reboot an Ardana cluster. Also added support for the new playbook
to rocketchat_notify role.

Use OPS-LM rather than ARD-SVC as the host group when looking up the
deployer node, as that is the actual name of the life cycle manager
group.

Update the manual/lib.sh shell library to support Ardana upgrade
workflow.

Add additional utility scripts to the manual area:
  * reboot-ardana.sh
    reboot an already running cloud using new ardana-reboot.yml.
  * test-ardana.sh
    run configured tempest and QA tests against an existing cloud.
  * update-ardana.sh
    update a running cloud to the configured update_to_cloudsource.
  * upgrade-ardana.sh
    upgrade a running cloud to the configured upgrade_cloudsource.

Add git ignorance for the ardana-bm-info area under the cloud/ansible
directory.